requirements.txt
langchain
langchain-community
langchain-openai
chromadb
openai
pypdf
tiktoken
streamlit
python-dotenv
langchain-experimental
rank_bm25
docx2txt
sentence-transformers